backend/requirements.txt
# FastAPI and server
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20

# Database
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
alembic==1.14.0

# Async task queue
celery[redis]==5.4.0
redis==5.2.1

# S3/MinIO client
boto3==1.35.86

# Document processing
pymupdf==1.25.1
python-docx==1.1.2

# AI
anthropic==0.42.0

# Utilities
pydantic==2.10.4
pydantic-settings==2.7.0
python-dotenv==1.0.1

# Testing
pytest==8.3.4
pytest-asyncio==0.25.0
httpx==0.28.1
